feat(table): apply theme and better performance
