+++ src/index.ts
@@ -1,15 +1,16 @@
+export * from './components/Chip';
+export * from './components/Card';
+export * from './components/Tabs';
+export * from './components/Radio';
 export * from './components/Theme';
 export * from './components/Modal';
-export * from './components/Button';
-export * from './components/TextInput';
-export * from './components/NumberInput/NumberInput';
-export * from './components/Checkbox';
-export * from './components/Chip';
-export * from './components/Select/Select';
-export * from './components/Card';
-export * from './components/Radio';
-export * from './components/Skeleton';
-export * from './components/Tabs';
 export * from './components/Table';
 export * from './components/Alert';
+export * from './components/Button';
+export * from './components/Popover';
+export * from './components/Checkbox';
+export * from './components/Skeleton';
+export * from './components/TextInput';
 export * from './components/ImageInput';
+export * from './components/Select/Select';
+export * from './components/NumberInput/NumberInput';
