+++ src/stories/components/Popover.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
-import { Popover } from "../../components/Popover/Popover";
+import { Popover } from "../../components/Popover";
 import { Button } from "../../components/Button";
 import { Edit, Delete } from "../../icons";
 
@@ -11,9 +11,11 @@ export default {
 } as ComponentMeta<typeof Popover>;
 
 const Template: ComponentStory<typeof Popover> = (args) => (
-    <Popover {...args}>
-        <Button size="medium" color="blue" label="Toggle Popover" />
-    </Popover>
+    <div style={{ width: '100%', height: '100vh', display: 'flex', alignItems: 'center', justifyContent: 'center'}}>
+        <Popover {...args}>
+            <Button size="medium" color="blue" label="Toggle Popover" />
+        </Popover>
+    </div>
 );
 
 export const Default = Template.bind({});
